Refactor: move linter to its own package
If we're going to add more linters, and more capabilities, it may help
to put it in its own package.
