+++ Makefile
@@ -45,10 +45,10 @@ lint:
 
 .PHONY: test
 test: envtest
-	KUBEBUILDER_ASSETS="$(shell $(ENVTEST) use $(ENVTEST_K8S_VERSION) -p path)" go test $(TESTARGS) -coverpkg=$(shell cat go.mod | head -1 | cut -d ' ' -f 2)/... ./...
+	KUBEBUILDER_ASSETS="$(shell $(ENVTEST) use $(ENVTEST_K8S_VERSION) -p path)" go test $(TESTARGS) ./...
 
 .PHONY: test-coverage
-test-coverage: TESTARGS = -cover -covermode=atomic -coverprofile=coverage.out
+test-coverage: TESTARGS = -cover -covermode=atomic -coverprofile=coverage.out -coverpkg=$(shell cat go.mod | head -1 | cut -d ' ' -f 2)/...
 test-coverage: test
 
 .PHONY: gosec-scan
