+++ pkg/lint/lint.go
@@ -1,4 +1,4 @@
-package templates
+package lint
 
 import (
 	"fmt"
